Move the json column mapper tests next to the config they prove
The wiring assertion reads global SessionFactoryOptions and has nothing to do
with Discovery, so it was reachable only through a file about the discovery
repository. Both tests extend BaseSpringBootTest and reuse its cached context,
leaving the signature baseline untouched.
